+++ benchmarks/allocation_size_benchmarks.cpp
@@ -8,7 +8,6 @@
 
 #include "umpire/Allocator.hpp"
 #include "umpire/ResourceManager.hpp"
-#include "umpire/resource/MemoryResourceTypes.hpp"
 
 #include "umpire/alloc/MallocAllocator.hpp"
 #if defined(UMPIRE_ENABLE_CUDA)
@@ -116,50 +115,8 @@ BENCHMARK_DEFINE_F(HipPinned, hipMallocHost)(benchmark::State& st) { allocation(
 BENCHMARK_DEFINE_F(HipPinned, hipFreeHost)(benchmark::State& st)   { deallocation(st); }
 #endif
 
-template <umpire::resource::MemoryResourceType Resource>
-class MemoryResourceAllocator : public AllocatorBenchmark
-{
-public:
-  using ::benchmark::Fixture::SetUp;
-  using ::benchmark::Fixture::TearDown;
-
-  MemoryResourceAllocator() : m_alloc{nullptr} {}
-
-  void SetUp(benchmark::State&) override final {
-    m_alloc = new umpire::Allocator{umpire::ResourceManager::getInstance().getAllocator(Resource)};
-  }
-
-  void TearDown(benchmark::State&) override final { delete m_alloc; }
-
-  virtual void* allocate(std::size_t nbytes) final { return m_alloc->allocate(nbytes); }
-  virtual void deallocate(void* ptr) final { m_alloc->deallocate(ptr); }
-private:
-  umpire::Allocator* m_alloc;
-};
-
-class HostResource : public MemoryResourceAllocator<umpire::resource::Host> {};
-BENCHMARK_DEFINE_F(HostResource, allocate)(benchmark::State& st) { allocation(st); }
-BENCHMARK_DEFINE_F(HostResource, deallocate)(benchmark::State& st) { deallocation(st); }
-
-#if defined(UMPIRE_ENABLE_DEVICE)
-class DeviceResource : public MemoryResourceAllocator<umpire::resource::Device> {};
-BENCHMARK_DEFINE_F(DeviceResource, allocate)(benchmark::State& st) { allocation(st); }
-BENCHMARK_DEFINE_F(DeviceResource, deallocate)(benchmark::State& st)   { deallocation(st); }
-
-class DevicePinnedResource : public MemoryResourceAllocator<umpire::resource::Pinned> {};
-BENCHMARK_DEFINE_F(DevicePinnedResource, allocate)(benchmark::State& st) { allocation(st); }
-BENCHMARK_DEFINE_F(DevicePinnedResource, deallocate)(benchmark::State& st) { deallocation(st); }
-#endif
-
-#if defined(UMPIRE_ENABLE_UM)
-class UnifiedResource : public MemoryResourceAllocator<umpire::resource::Unified> {};
-BENCHMARK_DEFINE_F(UnifiedResource, allocate)(benchmark::State& st) { allocation(st); }
-BENCHMARK_DEFINE_F(UnifiedResource, deallocate)(benchmark::State& st) { deallocation(st); }
-#endif
-
 ///////////////////////////////////////////////////////////////////////////////////
 // Register all the benchmarks
-// Base allocators
 BENCHMARK_REGISTER_F(Malloc, malloc)->RangeMultiplier(4)->Range(RangeLow, RangeHi);
 BENCHMARK_REGISTER_F(Malloc, free)->RangeMultiplier(4)->Range(RangeLow, RangeHi);
 
@@ -183,24 +140,6 @@ BENCHMARK_REGISTER_F(HipMallocManaged, hipFree)->RangeMultiplier(4)->Range(Range
 
 BENCHMARK_REGISTER_F(HipPinned, hipMallocHost)->RangeMultiplier(4)->Range(RangeLow, RangeHi);
 BENCHMARK_REGISTER_F(HipPinned, hipFreeHost)->RangeMultiplier(4)->Range(RangeLow, RangeHi);
-#endif
-
-///////////////////////////////////////////////////////////////////////////////////
-// Resources
-BENCHMARK_REGISTER_F(HostResource, allocate)->RangeMultiplier(4)->Range(RangeLow, RangeHi);
-BENCHMARK_REGISTER_F(HostResource, deallocate)->RangeMultiplier(4)->Range(RangeLow, RangeHi);
-
-#if defined(UMPIRE_ENABLE_DEVICE)
-BENCHMARK_REGISTER_F(DeviceResource, allocate)->RangeMultiplier(4)->Range(RangeLow, RangeHi);
-BENCHMARK_REGISTER_F(DeviceResource, deallocate)->RangeMultiplier(4)->Range(RangeLow, RangeHi);
-
-BENCHMARK_REGISTER_F(DevicePinnedResource, allocate)->RangeMultiplier(4)->Range(RangeLow, RangeHi);
-BENCHMARK_REGISTER_F(DevicePinnedResource, deallocate)->RangeMultiplier(4)->Range(RangeLow, RangeHi);
-#endif
-
-#if defined(UMPIRE_ENABLE_UM)
-BENCHMARK_REGISTER_F(UnifiedResource, allocate)->RangeMultiplier(4)->Range(RangeLow, RangeHi);
-BENCHMARK_REGISTER_F(UnifiedResource, deallocate)->RangeMultiplier(4)->Range(RangeLow, RangeHi);
 #endif
 
 BENCHMARK_MAIN();