+++ CHANGELOG.md
@@ -20,8 +20,8 @@ Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
 
 - Benchmark that overlooks overhead of malloc calls with No-Op memory resource.
 
-- Benchmark that measures the performance of different allocators and memory 
-  resource types.
+- Benchmark that measures the performance of different allocator types across
+  various sizes.
 
 ### Changed